pgsync-python/requirements.txt
psycopg2-binary
elasticsearch
elastic-transport
requests
